\name{sabaRbasics-package}
\alias{sabaRbasics-package}
\alias{sabaRbasics}
\docType{package}
\title{Basic Function I wish R had}
\description{for personal use}
\details{
\tabular{ll}{
Package: \tab sabaRbasics\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-08\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{}
